Create README.md
The garbage_search script for FiveM allows players to search through garbage bins, adding an immersive activity to the role-playing experience. This script includes animations for the searching process and provides configurable items and props that can be found. The language settings can be customized to either French or English through the config.lua file.